start definining interfaces
